refactor: new event
